Added Main Quest Explanation Dialogues, Clear Screen and Next Dialogue Functions
Added a function in the ProfessorOak class which consists of dialogues of Professor Oak and the player while explaining the main quest.

Added a function in the Player class which allows player to press enter after each dialogue in order to see the next dialogue, instead of printing all dialogues at once.

Added a function defined globally to clear the screen to prevent screen clutter.