Created UtilityFunctions header file and source file
Created UtilityFunctions header file in the Header Files folder. Created UtilityFunctions class in the header file. Declared the necessary functions as static for clearing screen and clearing buffer in the class.

Created the UtilityFunctions source file in the Source Files folder. Defined the functions declared in the UtilityFunctions header file, in the source file.