Created WildPokemonEncounterManager header and source files
Created the WildPokemonEncounterManager header file in the Header Files folder. Created the WildPokemonEncounterManager class in the header file. Declared the necessary function to implement retrieving random pokemon from the list of pokemon present in the vector of instance of Grass struct.

Created the WildPokemonEncounterManager source file in the Source Files folder. Defined the function declared in the header file, in the source file.

Made changes to Game source file by including the WildPokemonEncounterManager
header file and removing the initial cout statement from the case 1 in GameLoop function and replacing it with the code to get and display the wild pokemon encountered.

Declared and defined function to get the string format of PokemonType in the Pokemon header and source file respectively.