Created BattleManager header and source files and Implemented battle loop
Created BattleManager header file in the Header Files folder. Created the BattleManager class in the header file. Declared a function to implement the Battle Loop in the class.

Created the BattleManager source file in the Source Files folder. Included the necessary header files and defined the function to implement battle loop declared in the header file, in the source file.

Made changes to Pokemon header and source files by declaring and defining variable for max health and functions for attack, taking damage and checking whether pokemon has fainted.

Made changes to the Game source file by including the BattleManager header file and by creating an instance of BattleManager class and using the instance to call the battle loop function in case 1 of the Game Loop.