Created the BattleState header file
Created the BattleState header file in the Header Files folder. Created the BattleState struct in the header file. Declared the pointers of Pokemon class and variables to check for player turn and whether battle is still ongoing.

Made changes to BattleManager header file. Included the BattleState header file in the BattleManager header file. Created instance of BattleState struct. Added function to update battle state by checking whether any pokemon has fainted or not. Made changes to the Battle and BattleOutcome functions by removing their parameters.

Made changes to the BattleManager source file. Assigned the memory addresses of player pokemon and opponent pokemon instances to the Pokemon pointers of BattleState instance and set the values of player turn and battle ongoing boolean variables to true in the StartBattle function. Used the boolean variables and the Pokemon pointers of BattleState instance to check the state of the BattleLoop, Player turn and to call the Attack function. Defined the function to update battle state by using Pokemon pointers of BattleState instance to call the function  of Pokemon class to check whether Pokemon has fainted or not and then set the value of battle ongoing boolean variable accordingly. Made changes to BattleOutcome function definition by using Pokemon pointers of BattleState instance.

Made changes to the Pokemon header and source files by including the attack power variable in the parameterized constructor.

Made changes to the Player and Game source files by including the value of attack power for each pokemon during creation of Pokemon instance.