Created Bulbasaur, Charmander and Squirtle header and source files
Created the header and source files for Bulbasaur, Charmander and Squirtle in the Header Files and Source Files folders respectively. Created classes for each which inherit from the Pokemon class in the header files. Declared and defined necessary functions in header and source files.

Made changes to the Player source file by creating instances of Bulbasaur, Charmander and Squirtle classes in the function to choose pokemon and assigning the appropriate instance to player pokemon object in the switch case statement based on user choice.

Made changes to the Game source file by replacing the Pokemon class instances for forest grass with instances of Pidgy, Caterpie and Zubat classes.