Created header file for a struct for Pokemon Moves and implemented Pokemon move selection during battle
Created PokemonMove header file in the Pokemon folder in the Header files folder. Created a PokemonMove struct in the N_Pokemon namespace in the header file. Declared variables for name of the move and attack power of the move and defined parameterized constructor for initializing the variables in the struct.

Made changes to the Pokemon header file and source files. Included the PokemonMove header file, declared vector of instances of PokemonMove struct and other necessary variables that are used during the pokemon battle in the header file. Declared and defined functions for selecting move, using move and applying certain side effects of an opponent pokemon's move to the pokemon in the header and source files. Changed the parameterized constructor by replacing the variable for attack power with vector for instances of PokemonMove struct.

Made changes to all the header and source files for different pokemons. Removed the function for each pokemon's special attack, created and passed instances of PokemonMove struct with name and attack power for each as parameters to the parameterized contructor of Pokemon class, called the functions to select move and use move and implemented unique logic for Attack function in source files of each pokemon.

Made changes to the Game header and source files. Declared and defined functions for checking player pokemon's health before entering into battle and not allowing player to enter battle if health is 0 in the header and source files.

Made changes to BattleManager source file by including the function calls to check for opponent pokemon's moves' side effects and disable them after battle is over.