Implemented Paralyzed Effect in Battle Loop
Made changes to the IStatusEffect header file by removing the Pokemon header file inclusion and forward declaring the Pokemon class in N_Pokemon namespace in this header file.

Removed the logic for Paralyzed Effect in Bulbasaur source file and moved it to Paralyzed Effect source file.

Declared and defined variable and functions to get and set Paralyzed Effect in Pokemon header and source files respectively. Created pointer of IStatusEffect interface and declared and defined functions to create instance of classes implementing IStatusEffect interface (ParalyzedEffect class in this case), initialize the pointer of IStatusEffect interface and to call the functions of the implementing classes using the pointer.

Added the Paralyzed Effect checks in the BattleManager source file and called appropriate functions of Pokemon class which in turn call functions of ParalyzedEffect class.